Revert "moved countdown to semgroup package"
This reverts commit 85850ea7200010ad4f70dc1630878a4159a9c367.
