+++ src/main/java/nl/tudelft/model/CountdownController.java
@@ -1,6 +1,6 @@
-package nl.tudelft.semgroup4;
+package nl.tudelft.model;
 
-import nl.tudelft.model.Game;
+import nl.tudelft.semgroup4.Renderable;
 import nl.tudelft.semgroup4.resources.ResourcesWrapper;
 
 import org.newdawn.slick.Color;
